feat(theme): add shared theme tokens
