+++ ui-admin/theme/package.json
@@ -18,6 +18,7 @@
     "postpack": "rm -rf dist"
   },
   "dependencies": {
+    "@atls-ui/theme-tokens": "workspace:*",
     "rainbow-sprinkles": "1.0.0"
   },
   "devDependencies": {
