+++ ui-admin/theme/src/theme/dark/space.ts
@@ -1,15 +1,1 @@
-import type { SpaceTokens } from '../index.js'
-
-export const space: SpaceTokens = {
-  zero: '0',
-  g1: '1px',
-  g4: '4px',
-  g6: '6px',
-  g8: '8px',
-  g10: '10px',
-  g12: '12px',
-  g14: '14px',
-  g17: '17px',
-  g22: '22px',
-  g24: '24px',
-}
+export { space } from '@atls-ui/theme-tokens'
