+++ ui/theme-tokens/src/animations.ts
@@ -1,0 +1,17 @@
+export const animations = {
+  duration: {
+    instant: '0ms',
+    fast: '150ms',
+    normal: '200ms',
+    slow: '300ms',
+  },
+  easing: {
+    linear: 'linear',
+    ease: 'ease',
+    easeIn: 'ease-in',
+    easeOut: 'ease-out',
+    easeInOut: 'ease-in-out',
+  },
+} as const
+
+export type AnimationsTokens = typeof animations
