+++ ui/theme-tokens/src/opacities.ts
@@ -1,0 +1,7 @@
+export const opacities = {
+  transparent: 0,
+  disabled: 0.5,
+  visible: 1,
+} as const
+
+export type OpacitiesTokens = typeof opacities
