+++ ui/theme-tokens/src/radii.ts
@@ -1,0 +1,46 @@
+export const radii = {
+  none: 'none',
+  full: '100%',
+  f1: '1px',
+  f2: '2px',
+  f3: '3px',
+  f4: '4px',
+  f5: '5px',
+  f6: '6px',
+  f7: '7px',
+  f8: '8px',
+  f10: '10px',
+  f11: '11px',
+  f12: '12px',
+  f14: '14px',
+  f15: '15px',
+  f16: '16px',
+  f18: '18px',
+  f19: '19px',
+  f20: '20px',
+  f22: '22px',
+  f24: '24px',
+  f27: '27px',
+  f29: '29px',
+  f32: '32px',
+  f34: '34px',
+  f40: '40px',
+  f50: '50px',
+  f60: '60px',
+  f80: '80px',
+  f100: '100px',
+  f125: '125px',
+  f140: '140px',
+  f160: '160px',
+  f178: '178px',
+  f222: '222px',
+  f234: '234px',
+  f240: '240px',
+  f278: '278px',
+  f315: '315px',
+  f500: '500px',
+  f600: '600px',
+  f9999: '9999px',
+} as const
+
+export type RadiiTokens = typeof radii
