+++ ui/theme-tokens/src/space.ts
@@ -1,0 +1,17 @@
+export const space = {
+  zero: '0',
+  g1: '1px',
+  g4: '4px',
+  g6: '6px',
+  g8: '8px',
+  g10: '10px',
+  g12: '12px',
+  g14: '14px',
+  g16: '16px',
+  g17: '17px',
+  g22: '22px',
+  g24: '24px',
+  g32: '32px',
+} as const
+
+export type SpaceTokens = typeof space
